Fixed bug where BASE_PORT is appended to instead of incremented by one
When you use A SystemError occurred.

Debug Information:
    Node Version: v6.11.2
    Ghost-CLI Version: 1.1.0
    Environment: production
    Command: 'ghost install'

Please refer to https://docs.ghost.org/docs/troubleshooting#section-cli-errors for troubleshooting. and specify a blog url without a port number, BASE_PORT is treated as a string so by the third installation it fails as it tried port 2368, then 23681, then 236811 (which is greater than 65535).