chore(systemd): ignore error in `get-uid` and return null
refs #587

Instead of returning an error in the `get-uid` util we can simplify it and just return null.

Explanation:

Possible reasons for error:
- Command is not known by OS or
- User doesn't exist

The first case is not very likely to happen, as systemd is never setup on a Windows. And the second case would exactly be the result we want. false if the user doesn't exist.