fix(run): Don't set contentPath in config when running
This was added in the past to ensure contentPath will be valid for ghost. Given the contentPath is set (if not already) in the migrate step, this extra check causes invalid manually edited configuration to be overwritten with a (basically) empty file. Since a lot of people test their config by restarting ghost, this hurts more people than it helps