fix(migrations): sudo - don't fail with existing settings folder
issue #776

In production sites, when the ghost user is used, the `content/*` folder needs to be owned by the Ghost user so the ghost application will not run into permission issues. In the migration, to make sure this happens, we run the `mkdir` command using sudo (i.e. `sudo -E -u ghost mkdir {path}`). When the Ghost user isn't needed, we use the `fs-extra` library to make sure the directory exists (specifically `ensureDirSync`). `fs.ensureDirSync` is equivilant to running `mkdir` with the `p` flag, so the behaviour between sudo and no-sudo differs. By adding the `p` flag to the sudo command, the behaviour between the two cases are equalized, and the migrations won't fail if the content/settings folder exists when the Ghost user is used.