Ensure X-Forwarded-For is correct in default case
refs: https://github.com/orgs/TryGhost/projects/57/views/21

X-Forwarded-For is used to determine the client IP address within Ghost, and is used in some rate-limiting scenarios.
The original setting for this was to use Nginx's $proxy_add_x_forwarded_for variable, which appends to the list.

Since it's trivial to spoof a client address with this mode, the default case should be to assume this is the most external proxy (i.e. this Nginx server is what clients connect to, with no other proxies in-between).

If someone has a setup which uses a CDN or additional reverse proxies, then this configuration will need changing to use $proxy_add_x_forwarded_for, and additional configuration may be required to ensure that only the next proxy in the chain can access Nginx.